% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascade_functions.R
\name{predict_spline_cascade}
\alias{predict_spline_cascade}
\title{Predict from a cascading spline model}
\usage{
predict_spline_cascade(fit, newdata)
}
\arguments{
\item{fit}{the output of a \code{run_spline_cascade} function call, or otherwise a list that
identifies where on the file system to find the saved cascade model,
e.g. \code{list(working_dir = "/path/to/output_dir/model_label/")}}

\item{newdata}{a data frame with covariate values to use for prediction; must also include the
variables passed to \code{stage_id_vars} in \code{run_spline_cascade},
(e.g. \code{super_region_id} and \code{location_id}). To predict for levels higher in the
hierarchy, pass in \code{NA} values to the stage ID variables lower in the cascade}
}
\value{
data frame with predictions and an ID that identifies which model in the cascade
made the prediction
}
\description{
\code{predict_spline_cascade} uses a model fit by \code{run_spline_cascade} to make predictions
}
\examples{
require(dplyr)
df_pred <- expand.grid(
  stringsAsFactors = FALSE,
  age_start = seq(0, 100, by = 10),
  location_id = c(NA, "102")) \%>\%
    mutate(
      age_end = age_start,
      data_id = 1:nrow(.),
      super_region_name = "High-income"
    )

preds1 <- predict_spline_cascade(
  fit = fit1,
  newdata = df_pred
)

}
