% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascade_functions.R
\name{run_spline_cascade}
\alias{run_spline_cascade}
\title{Fit a cascading spline model}
\usage{
run_spline_cascade(
  stage1_model_object,
  df,
  col_obs,
  col_obs_se,
  col_study_id,
  stage_id_vars,
  thetas,
  output_dir,
  model_label,
  inner_print_level = 2L,
  inner_max_iter = 1000L,
  overwrite_previous = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{stage1_model_object}{a MR-BRT model object (post-fitting) that includes one spline}

\item{df}{a data frame with all needed variables for running the model}

\item{col_obs}{column name for the dependent variable}

\item{col_obs_se}{column name for the standard error of the dependent variable}

\item{col_study_id}{column name for the variable with unique group IDs, for random effects}

\item{stage_id_vars}{a vector of column names identifying cascade stages
(e.g. \code{c(super_region_id, region_id, location_id)});
must have a nested structure and be in hierarchical order from left to right}

\item{thetas}{multiplier for the standard deviations of the priors on the spline betas;
a vector with an element for each varialbe in \code{stage_id_vars}}

\item{output_dir}{path to a directory in which the working directory will be created}

\item{model_label}{name of the model; will also be the name of the working directory}

\item{inner_print_level}{integer indicating the amount of information to be displayed by IPOPT
during the fitting process; e.g. \code{1L} or \code{5L}}

\item{inner_max_iter}{maximum number of iterations allowed per model}

\item{overwrite_previous}{logical; whether to overwrite pre-existing results
at \code{file.path(output_dir, model_label)}}
}
\value{
list of model info, including where the working directory with saved outputs is located
}
\description{
\code{run_spline_cascade} implements hierarchical spline estimation, using the estimated spline
coefficients from a parent model (e.g. global) as Bayesian priors in child models
(e.g. for each region)
}
\examples{
fit1 <- run_spline_cascade(
  stage1_model_object = mod3,
  df = df,
  col_obs = "logit_mort",
  col_obs_se = "logit_mort_se",
  col_study_id = "location_id",
  stage_id_vars = c("super_region_name", "location_id"),
  thetas = c(1,1),
  output_dir = "FILEPATH",
  model_label = "mrbrt_cascade_test2",
  overwrite_previous = TRUE
)

}
